Fix styling for GitHub Actions

+++ test/_06_NuclearSecurity.t.sol
@@ -4,13 +4,12 @@ pragma solidity ^0.8.24;
 import {Test} from "forge-std/Test.sol";
 import {AxilProtocolV1} from "../src/AxilProtocolV1.sol";
 
-
 /**
  * @title Nuclear Security Tests
  * @author Axil Protocol Team
  * @notice Extreme security validation with 1B MON attack simulations
  * @dev Tests all possible attack vectors with maximum amounts
- * 
+ *
  * ╔══════════════════════════════════════════════════════════╗
  * ║              NUCLEAR SECURITY AUDIT REPORT               ║
  * ╠══════════════════════════════════════════════════════════╣
@@ -29,11 +28,11 @@ import {AxilProtocolV1} from "../src/AxilProtocolV1.sol";
  */
 contract NuclearSecurityTest is Test {
     AxilProtocolV1 public axil;
-    
+
     // Constants
     uint256 constant MAX_MON_SUPPLY = 1_000_000_000 ether; // 1 Billion MON
     uint256 constant SIGNER_KEY = 0xA1;
-    
+
     // Test addresses
     address public signer;
     address public agent = address(0x8);
@@ -52,16 +51,9 @@ contract NuclearSecurityTest is Test {
     function setUp() public {
         signer = vm.addr(SIGNER_KEY);
         vm.startPrank(address(0x1));
-        axil = new AxilProtocolV1(
-            address(0x1), 
-            signer, 
-            address(0x3), 
-            address(0x4), 
-            address(0x5), 
-            keccak256("SALT")
-        );
+        axil = new AxilProtocolV1(address(0x1), signer, address(0x3), address(0x4), address(0x5), keccak256("SALT"));
         vm.stopPrank();
-        
+
         // Fund all addresses with 1B MON each
         vm.deal(agent, MAX_MON_SUPPLY);
         vm.deal(attacker, MAX_MON_SUPPLY);
@@ -72,13 +64,15 @@ contract NuclearSecurityTest is Test {
      * @dev Helper to generate domain separator
      */
     function _getDomainSeparator() internal view returns (bytes32) {
-        return keccak256(abi.encode(
-            keccak256("EIP712Domain(string name,string version,uint256 chainId,address verifyingContract)"),
-            keccak256(bytes("AxilProtocolV1")),
-            keccak256(bytes("1")),
-            block.chainid,
-            address(axil)
-        ));
+        return keccak256(
+            abi.encode(
+                keccak256("EIP712Domain(string name,string version,uint256 chainId,address verifyingContract)"),
+                keccak256(bytes("AxilProtocolV1")),
+                keccak256(bytes("1")),
+                block.chainid,
+                address(axil)
+            )
+        );
     }
 
     /**
@@ -93,9 +87,9 @@ contract NuclearSecurityTest is Test {
         uint128 _salt,
         address _agent
     ) internal view returns (bytes memory) {
-        bytes32 structHash = keccak256(abi.encode(
-            EXECUTE_TYPEHASH, _merchant, _user, _packedIntent, _amount, _deadline, _salt, _agent
-        ));
+        bytes32 structHash = keccak256(
+            abi.encode(EXECUTE_TYPEHASH, _merchant, _user, _packedIntent, _amount, _deadline, _salt, _agent)
+        );
         bytes32 finalHash = keccak256(abi.encodePacked("\x19\x01", _getDomainSeparator(), structHash));
         (uint8 v, bytes32 r, bytes32 s) = vm.sign(SIGNER_KEY, finalHash);
         return abi.encodePacked(r, s, v);
@@ -103,12 +97,11 @@ contract NuclearSecurityTest is Test {
 
     /**
      * @notice Test 6.1: Maximum supply overflow protection
-
-
-* @dev Attempt to process 1B MON transaction
+     *
+     *
+     * @dev Attempt to process 1B MON transaction
      */
     function test_Attack_MaxSupplyOverflow() public {
-       
         // forge-lint: disable-next-line(unsafe-typecast)
         uint128 amount = uint128(MAX_MON_SUPPLY);
         uint256 deadline = block.timestamp + 1000;
@@ -120,7 +113,7 @@ contract NuclearSecurityTest is Test {
 
         vm.prank(agent);
         axil.execute{value: amount}(merchant, user, packedIntent, deadline, salt, signature);
-        
+
         assertTrue(axil.isIntentExecuted(packedIntent));
         emit SecurityBreachAttempt("MAX_UINT128_ATTACK", true);
     }
@@ -135,26 +128,22 @@ contract NuclearSecurityTest is Test {
         bytes32 packedIntent = axil.packIntent(2, 2);
 
         ReentrancyAttacker attackerContract = new ReentrancyAttacker(axil);
-        bytes memory signature = _createSignature(merchant, user, packedIntent, amount, deadline, salt, address(attackerContract));
+        bytes memory signature =
+            _createSignature(merchant, user, packedIntent, amount, deadline, salt, address(attackerContract));
 
         vm.deal(address(attackerContract), amount);
         vm.prank(address(attackerContract));
-        
+
         attackerContract.attack(merchant, user, packedIntent, deadline, salt, signature, amount);
-        
-        (bool success, ) = address(axil).call(
-            abi.encodeWithSelector(
-                axil.execute.selector,
-                address(0),
-                address(0),
-                bytes32(0),
-                block.timestamp,
-                0,
-                ""
-            )
-        );
+
+        (bool success,) = address(axil)
+            .call(
+                abi.encodeWithSelector(
+                    axil.execute.selector, address(0), address(0), bytes32(0), block.timestamp, 0, ""
+                )
+            );
         assertFalse(success, "Reentrancy blocked successfully");
-        
+
         emit SecurityBreachAttempt("REENTRANCY_ATTACK", true);
     }
 
@@ -167,19 +156,18 @@ contract NuclearSecurityTest is Test {
         uint128 salt = 98765;
         bytes32 packedIntent = axil.packIntent(3, 3);
 
-        bytes32 structHash = keccak256(abi.encode(
-            EXECUTE_TYPEHASH, merchant, user, packedIntent, amount, deadline, salt, agent
-        ));
+        bytes32 structHash =
+            keccak256(abi.encode(EXECUTE_TYPEHASH, merchant, user, packedIntent, amount, deadline, salt, agent));
         bytes32 finalHash = keccak256(abi.encodePacked("\x19\x01", _getDomainSeparator(), structHash));
         (uint8 v, bytes32 r, bytes32 s) = vm.sign(SIGNER_KEY, finalHash);
-        
+
         bytes32 sModified = bytes32(uint256(s) ^ 1);
         bytes memory maliciousSignature = abi.encodePacked(r, sModified, v);
 
         vm.prank(agent);
         vm.expectRevert(AxilProtocolV1.Axil__InvalidSignature.selector);
         axil.execute{value: amount}(merchant, user, packedIntent, deadline, salt, maliciousSignature);
-        
+
         emit SecurityBreachAttempt("SIGNATURE_MALLEABILITY", true);
     }
 
@@ -199,7 +187,7 @@ contract NuclearSecurityTest is Test {
         vm.prank(agent);
         vm.expectRevert(AxilProtocolV1.Axil__SignatureExpired.selector);
         axil.execute{value: amount}(merchant, user, packedIntent, originalDeadline, salt, signature);
-        
+
         emit SecurityBreachAttempt("DEADLINE_MANIPULATION", true);
     }
 
@@ -207,20 +195,19 @@ contract NuclearSecurityTest is Test {
      * @notice Test 6.5: Integer overflow protection
      */
     function test_Attack_IntegerOverflow() public {
-       
-       // forge-lint: disable-next-line(unsafe-typecast)
+        // forge-lint: disable-next-line(unsafe-typecast)
         uint128 amount = uint128(MAX_MON_SUPPLY);
         vm.deal(agent, amount);
-        
+
         uint256 deadline = block.timestamp + 1000;
         uint128 salt = 22222;
         bytes32 packedIntent = axil.packIntent(5, 5);
 
-bytes memory signature = _createSignature(merchant, user, packedIntent, amount, deadline, salt, agent);
+        bytes memory signature = _createSignature(merchant, user, packedIntent, amount, deadline, salt, agent);
 
         vm.prank(agent);
         axil.execute{value: amount}(merchant, user, packedIntent, deadline, salt, signature);
-        
+
         assertTrue(axil.isIntentExecuted(packedIntent));
         emit SecurityBreachAttempt("INTEGER_OVERFLOW", true);
     }
@@ -230,15 +217,8 @@ bytes memory signature = _createSignature(merchant, user, packedIntent, amount, 
      */
     function test_Attack_ZeroAddress() public {
         vm.expectRevert(AxilProtocolV1.Axil__ZeroAddressNotAllowed.selector);
-        new AxilProtocolV1(
-            address(0), 
-            signer, 
-            address(0x3), 
-            address(0x4), 
-            address(0x5), 
-            keccak256("SALT")
-        );
-        
+        new AxilProtocolV1(address(0), signer, address(0x3), address(0x4), address(0x5), keccak256("SALT"));
+
         emit SecurityBreachAttempt("ZERO_ADDRESS_ATTACK", true);
     }
 
@@ -255,13 +235,13 @@ bytes memory signature = _createSignature(merchant, user, packedIntent, amount, 
 
         vm.prank(agent);
         axil.execute{value: amount}(merchant, user, packedIntent, deadline, salt, signature);
-        
+
         vm.chainId(31337);
-        
+
         vm.prank(agent);
         vm.expectRevert(AxilProtocolV1.Axil__IntentAlreadyExecuted.selector);
         axil.execute{value: amount}(merchant, user, packedIntent, deadline, salt, signature);
-        
+
         emit SecurityBreachAttempt("CROSS_CHAIN_REPLAY", true);
     }
 
@@ -271,40 +251,34 @@ bytes memory signature = _createSignature(merchant, user, packedIntent, amount, 
     function test_Attack_GasGriefing() public {
         uint128 amount = 1 ether;
         uint256 deadline = block.timestamp + 1000;
-        
+
         uint256 gasStart = gasleft();
-        
-        for(uint i = 1; i <= 100; i++) {
+
+        for (uint256 i = 1; i <= 100; i++) {
             // forge-lint: disable-next-line(unsafe-typecast)
             uint128 salt = uint128(i);
             // forge-lint: disable-next-line(unsafe-typecast)
             bytes32 packedIntent = axil.packIntent(8, uint128(i));
-            
+
             bytes memory invalidSignature = abi.encodePacked(
-                bytes32(uint256(i)), 
-                bytes32(uint256(i + 1)), 
+                bytes32(uint256(i)),
+                bytes32(uint256(i + 1)),
                 // forge-lint: disable-next-line(unsafe-typecast)
                 uint8(i % 256)
             );
 
             vm.prank(attacker);
-            (bool success, ) = address(axil).call{value: amount}(
+            (bool success,) = address(axil).call{value: amount}(
                 abi.encodeWithSelector(
-                    axil.execute.selector,
-                    merchant, 
-                    user, 
-                    packedIntent, 
-                    deadline, 
-                    salt, 
-                    invalidSignature
+                    axil.execute.selector, merchant, user, packedIntent, deadline, salt, invalidSignature
                 )
             );
             assertFalse(success);
         }
-        
+
         uint256 gasUsed = gasStart - gasleft();
         assertTrue(gasUsed < 5_000_000, "Gas griefing blocked successfully");
-        
+
         emit SecurityBreachAttempt("GAS_GRIEFING", true);
     }
 
@@ -313,13 +287,12 @@ bytes memory signature = _createSignature(merchant, user, packedIntent, amount, 
      * @dev Attempt 1B MON flash loan attack
      */
     function test_Attack_FlashLoanBillionMon() public {
-        
         // forge-lint: disable-next-line(unsafe-typecast)
         uint128 flashLoanAmount = uint128(MAX_MON_SUPPLY);
-        
+
         vm.startPrank(hacker);
         vm.deal(hacker, flashLoanAmount);
-        
+
         uint256 deadline = block.timestamp + 1;
         uint128 salt = 99999;
         bytes32 packedIntent = axil.packIntent(9, 9);
@@ -328,35 +301,35 @@ bytes memory signature = _createSignature(merchant, user, packedIntent, amount, 
 
         vm.expectRevert(AxilProtocolV1.Axil__DeadlineTooShort.selector);
         axil.execute{value: flashLoanAmount}(hacker, hacker, packedIntent, deadline, salt, signature);
-        
+
         vm.stopPrank();
-        
+
         emit SecurityBreachAttempt("FLASH_LOAN_1B_ATTACK", true);
     }
 
-/**
+    /**
      * @notice Test 6.10: Complete system integrity verification
      * @dev Verify all 10 unique intents work correctly
      */
     function test_SystemIntegrity() public {
         uint128 amount = 100 ether;
         uint256 deadline = block.timestamp + 1000;
-        
+
         // Generate and execute 10 unique intents - each only ONCE
-        for(uint i = 1; i <= 10; i++) {
+        for (uint256 i = 1; i <= 10; i++) {
             // forge-lint: disable-next-line(unsafe-typecast)
             uint128 salt = uint128(i * 100000);
             // forge-lint: disable-next-line(unsafe-typecast)
             bytes32 packedIntent = axil.packIntent(10, uint128(i));
-            
+
             bytes memory signature = _createSignature(merchant, user, packedIntent, amount, deadline, salt, agent);
 
             vm.prank(agent);
             axil.execute{value: amount}(merchant, user, packedIntent, deadline, salt, signature);
-            
+
             assertTrue(axil.isIntentExecuted(packedIntent));
         }
-        
+
         emit SecurityBreachAttempt("SYSTEM_INTEGRITY_CHECK", true);
     }
 
@@ -367,13 +340,13 @@ bytes memory signature = _createSignature(merchant, user, packedIntent, amount, 
     function testFuzz_BillionMonAttack(uint128 amount, uint128 salt, uint256 timestamp) public {
         // Bound values for realistic testing up to 1B MON
         amount = uint128(bound(amount, 0.001 ether, MAX_MON_SUPPLY));
-        
+
         // Ensure salt is not zero (mask cannot be 0)
         if (salt == 0) salt = 1;
         salt = uint128(bound(salt, 1, type(uint128).max));
-        
+
         timestamp = bound(timestamp, block.timestamp + 1 hours, block.timestamp + 30 days);
-        
+
         uint256 deadline = timestamp;
         bytes32 packedIntent = keccak256(abi.encodePacked(salt, "BILLION_ATTACK"));
 
@@ -381,7 +354,7 @@ bytes memory signature = _createSignature(merchant, user, packedIntent, amount, 
 
         vm.deal(agent, amount);
         vm.prank(agent);
-        
+
         // Single execution - no replay test
         axil.execute{value: amount}(merchant, user, packedIntent, deadline, salt, signature);
         assertTrue(axil.isIntentExecuted(packedIntent));
@@ -394,11 +367,11 @@ bytes memory signature = _createSignature(merchant, user, packedIntent, amount, 
 contract ReentrancyAttacker {
     AxilProtocolV1 public axil;
     bool public attacked;
-    
+
     constructor(AxilProtocolV1 _axil) {
         axil = _axil;
     }
-    
+
     function attack(
         address merchant,
         address user,
@@ -410,21 +383,16 @@ contract ReentrancyAttacker {
     ) external payable {
         axil.execute{value: amount}(merchant, user, packedIntent, deadline, salt, signature);
     }
-    
+
     receive() external payable {
         if (!attacked) {
             attacked = true;
-            (bool success, ) = address(axil).call(
-                abi.encodeWithSelector(
-                    axil.execute.selector,
-                    address(0),
-                    address(0),
-                    bytes32(0),
-                    block.timestamp,
-                    0,
-                    ""
-                )
-            );
+            (bool success,) = address(axil)
+                .call(
+                    abi.encodeWithSelector(
+                        axil.execute.selector, address(0), address(0), bytes32(0), block.timestamp, 0, ""
+                    )
+                );
             require(!success, "Reentrancy should fail");
         }
     }
